Audit artifact management changes:
Logic update:
All transient artifacts are stored in different databases, grouped by processes. Cannot keep databases for all processes open at the same time because of file descriptor or other system resources contention. So, keeping a list of currently open databases, and list of existing (and closed) list of databases. Whenever a process operates on a transient artifact the appropriate database is looked for in the opened databases. If found then we are done. If not found then the database is looked for in the closed list of databases. If found in closed list then we try to open that, move that to opened list of databases and we are done. If not found in closed either then it means that it is a new one then we create a new one and put that in the list of opened databases.
In the steps mentioned above, because of system resources contention, (re)opening of a database might fail. To resolve that we close existing databases until able to (re)open the required one.

1) By default using 'LevelDBStore' for artifacts databases. High overhead in case of 'BerkeleyDB'.
2) 'maxopenretries' key added in 'spade.reporter.audit.artifact.ArtifactManager.config'. Used to specify how many times to try (re)opening a transient artifact map.
3) Modified external memory map creation API modified to control: 1) appending current time to db name for map, and 2) delete db dir if it exists.
4) Updated external store API to add functionality for closing database, reopening the database, checking if database is closed.
5) Added 'InMemoryStore' (alternative to 'LevelDBStore', and 'BerkeleyDB'), which keeps everything in memory instead of disk.